+++ src/main/java/org/betonquest/betonquest/quest/event/spawn/SpawnMobEventFactory.java
@@ -19,6 +19,7 @@ import org.betonquest.betonquest.quest.event.PrimaryServerThreadStaticEvent;
 import org.betonquest.betonquest.quest.registry.processor.VariableProcessor;
 import org.betonquest.betonquest.utils.Utils;
 import org.bukkit.entity.EntityType;
+import org.bukkit.entity.Mob;
 import org.jetbrains.annotations.Nullable;
 
 /**
@@ -48,24 +49,27 @@ public class SpawnMobEventFactory implements EventFactory, StaticEventFactory {
 
     @Override
     public Event parseEvent(final Instruction instruction) throws InstructionParseException {
-        return new PrimaryServerThreadEvent(createSpawnEntityEvent(instruction), data);
+        return new PrimaryServerThreadEvent(createSpawnMobEvent(instruction), data);
     }
 
     @Override
     public StaticEvent parseStaticEvent(final Instruction instruction) throws InstructionParseException {
-        return new PrimaryServerThreadStaticEvent(createSpawnEntityEvent(instruction), data);
+        return new PrimaryServerThreadStaticEvent(createSpawnMobEvent(instruction), data);
     }
 
     /**
-     * Creates a new spawn entity event from the given instruction.
+     * Creates a new spawn mob event from the given instruction.
      *
      * @param instruction the instruction to create the event from
      * @return the created event
      * @throws InstructionParseException if the instruction could not be parsed
      */
-    public NullableEventAdapter createSpawnEntityEvent(final Instruction instruction) throws InstructionParseException {
+    public NullableEventAdapter createSpawnMobEvent(final Instruction instruction) throws InstructionParseException {
         final VariableLocation loc = instruction.getLocation();
         final EntityType type = instruction.getEntity();
+        if (type.getEntityClass() == null || !Mob.class.isAssignableFrom(type.getEntityClass())) {
+            throw new InstructionParseException("The entity type must be a mob");
+        }
         final VariableNumber amount = instruction.getVarNum();
         final String nameString = instruction.getOptional("name");
         final VariableString name = nameString == null ? null : new VariableString(variableProcessor,